[RENAME] 포인트 도메인 예외처리 디렉토리 이동
