+++ src/main/java/kr/hhplus/be/server/domain/point/exception/ExceedsMaximumPointException.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.exception;
+package kr.hhplus.be.server.domain.point.exception;
 
 public class ExceedsMaximumPointException extends RuntimeException {
     public ExceedsMaximumPointException(String message) {
