+++ src/main/java/kr/hhplus/be/server/domain/point/exception/NotFoundUserException.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.exception;
+package kr.hhplus.be.server.domain.point.exception;
 
 public class NotFoundUserException extends RuntimeException {
     public NotFoundUserException(String message) {
