+++ src/main/java/kr/hhplus/be/server/domain/point/service/PointService.java
@@ -1,6 +1,6 @@
 package kr.hhplus.be.server.domain.point.service;
 
-import kr.hhplus.be.server.common.exception.NotFoundUserException;
+import kr.hhplus.be.server.domain.point.exception.NotFoundUserException;
 import kr.hhplus.be.server.domain.point.dto.response.PointResponse;
 import kr.hhplus.be.server.domain.point.entity.Point;
 import kr.hhplus.be.server.domain.point.repository.PointRepository;
